move asm parsing code into `rustc_parse`
